requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2